+++ @planetarium/tx/vitest.config.ts
@@ -2,5 +2,6 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
+    cache: false
   },
 });