+++ Libplanet.Net.Tests/Consensus/GossipTest.cs
@@ -52,7 +52,7 @@ namespace Libplanet.Net.Tests.Consensus
             var gossip1 = CreateGossip(
                 message =>
                 {
-                    if (message is ConsensusProposalMsg)
+                    if (message.Content is ConsensusProposalMsg)
                     {
                         received1 = true;
                     }
@@ -63,7 +63,7 @@ namespace Libplanet.Net.Tests.Consensus
             var gossip2 = CreateGossip(
                 message =>
                 {
-                    if (message is ConsensusProposalMsg)
+                    if (message.Content is ConsensusProposalMsg)
                     {
                         received2 = true;
                         receivedEvent.Set();
@@ -105,7 +105,7 @@ namespace Libplanet.Net.Tests.Consensus
             var gossip1 = CreateGossip(
                 message =>
                 {
-                    if (message is ConsensusProposalMsg)
+                    if (message.Content is ConsensusProposalMsg)
                     {
                         received1++;
                     }
@@ -116,7 +116,7 @@ namespace Libplanet.Net.Tests.Consensus
             var gossip2 = CreateGossip(
                 message =>
                 {
-                    if (message is ConsensusProposalMsg)
+                    if (message.Content is ConsensusProposalMsg)
                     {
                         received2++;
                     }