+++ Libplanet.Net/Messages/ConsensusProposalMsg.cs
@@ -39,11 +39,11 @@ namespace Libplanet.Net.Messages
         /// </summary>
         public Proposal Proposal { get; }
 
-        /// <inheritdoc cref="Message.DataFrames"/>
+        /// <inheritdoc cref="MessageContent.DataFrames"/>
         public override IEnumerable<byte[]> DataFrames =>
             new List<byte[]> { Proposal.ToByteArray() };
 
-        /// <inheritdoc cref="Message.MessageType"/>
+        /// <inheritdoc cref="MessageContent.MessageType"/>
         public override MessageType Type => MessageType.ConsensusProposal;
 
         public override bool Equals(ConsensusMsg? other)