+++ Libplanet/Tx/InvalidTxNonceException.cs
@@ -1,13 +1,14 @@
 using System;
-using System.Diagnostics.CodeAnalysis;
 using Libplanet.Blockchain;
+using Libplanet.Blocks;
 
 namespace Libplanet.Tx
 {
     /// <summary>
-    /// The exception that is thrown when the <see cref="Transaction{T}.Nonce"/>
-    /// is different from <see cref="BlockChain{T}.GetNextTxNonce"/> result of
-    /// the <see cref="Transaction{T}.Signer"/>.
+    /// An exception that is thrown when the <see cref="Transaction{T}.Nonce"/>
+    /// of a <see cref="Transaction{T}"/> included in a <see cref="Block{T}"/>
+    /// is inconsistent with the expected nonce for the <see cref="Transaction{T}.Signer"/>
+    /// when appending to a <see cref="BlockChain{T}"/>.
     /// </summary>
     [Serializable]
     public sealed class InvalidTxNonceException : InvalidTxException
@@ -17,35 +18,27 @@ namespace Libplanet.Tx
         /// <see cref="InvalidTxNonceException"/> class.
         /// </summary>
         /// <param name="message">The message that describes the error.</param>
-        /// <param name="txid">The invalid <see cref="Transaction{T}"/>'s
+        /// <param name="txId">The invalid <see cref="Transaction{T}"/>'s
         /// <see cref="Transaction{T}.Id"/>.  It is automatically included to
         /// the <see cref="Exception.Message"/> string.</param>
-        /// <param name="expectedNonce"><see cref="BlockChain{T}.GetNextTxNonce"/>
-        /// result of the <see cref="Transaction{T}.Signer"/>.</param>
-        /// <param name="improperNonce">The actual
-        /// <see cref="Transaction{T}.Nonce"/>.</param>
-        [SuppressMessage(
-            "Microsoft.StyleCop.CSharp.ReadabilityRules",
-            "SA1118",
-            Justification = "A long error message should be multiline.")]
+        /// <param name="expectedNonce">The expected <see cref="Transaction{T}.Nonce"/> value
+        /// for the <see cref="Transaction{T}"/>.</param>
+        /// <param name="improperNonce">The actual <see cref="Transaction{T}.Nonce"/>.</param>
         public InvalidTxNonceException(
             string message,
-            TxId txid,
+            TxId txId,
             long expectedNonce,
             long improperNonce)
-            : base(
-                $"{message}\n" +
-                $"Expected nonce: {expectedNonce}\n" +
-                $"Improper nonce: {improperNonce}",
-                txid)
+            : base(message, txId)
         {
             ExpectedNonce = expectedNonce;
             ImproperNonce = improperNonce;
         }
 
         /// <summary>
-        /// <see cref="BlockChain{T}.GetNextTxNonce"/> result of the
-        /// <see cref="Transaction{T}.Signer"/>.
+        /// The expected <see cref="Transaction{T}.Nonce"/> value
+        /// for the <see cref="Transaction{T}"/> with <see cref="TxId"/>
+        /// as its <see cref="Transaction{T}.Id"/>.
         /// </summary>
         public long ExpectedNonce { get; }